refactor: consolidate model logo imports to a single configuration file
